+++ PetRescue/PetRescue.Data/Models/Pet.cs
@@ -6,7 +6,7 @@ namespace PetRescue.Data.Models
     public partial class Pet
     {
         public Guid PetId { get; set; }
-        public int PetStatus { get; set; }
+        public byte PetStatus { get; set; }
         public Guid CenterId { get; set; }
         public Guid InsertedBy { get; set; }
         public DateTime InsertedAt { get; set; }
